requirements.txt
openai
pygithub
